+++ src/autoComplete/index.tsx
@@ -43,8 +43,8 @@ interface AutoCompleteProps {
   readonly buttonTitle?: string,
   readonly showList?: boolean,
   readonly valid?: boolean,
-  readonly onCheckingEnd?: () => void,
-} 
+  readonly onDoneAnimationEnd?: () => void,
+}
 
 interface AutoCompleteState {
   readonly busy: boolean,
@@ -256,7 +256,7 @@ export default class AutoComplete extends Component<AutoCompleteProps, AutoCompl
           loadingItemIndex={this.props.loadingItemIndex}
           itemClassName={this.props.itemClassName}
           valid={this.props.valid}
-          onCheckingEnd={this.props.onCheckingEnd}
+          onDoneAnimationEnd={this.props.onDoneAnimationEnd}
         />
         <style jsx>{style}</style>
       </div>