+++ src/loader/index.unit.tsx
@@ -1,6 +1,9 @@
 import React from 'react'
 import Loader from 'loader'
 import CircleIcon from 'icon/circleIcon'
+import CheckIcon from 'icon/checkIcon'
+
+jest.useFakeTimers()
 
 describe('Loader', () => {
   it('Should have a custom className', () => {
@@ -18,5 +21,19 @@ describe('Loader', () => {
   it('Should be inlined', () => {
     const wrapper = mount(<Loader inline />)
     expect(wrapper.prop('inline')).toBe(true)
+  })
+
+  it('Should show the done icon', () => {
+    const wrapper = mount(<Loader done />)
+    expect(wrapper.find(CheckIcon)).toHaveLength(1)
+  })
+
+  it('Should fire the callback event when done', () => {
+    const event = jest.fn()
+    const wrapper = mount(<Loader onDoneAnimationEnd={event} />)
+    wrapper.setProps({ done: true })
+    expect(event).not.toBeCalled()
+    jest.advanceTimersByTime(1500)
+    expect(event).toBeCalled()
   })
 })