+++ src/loader/story.tsx
@@ -9,12 +9,18 @@ import Loader from 'loader'
 const stories = storiesOf('Loader', module)
 stories.addDecorator(withKnobs)
 
+const onAnimationEnd = () => {
+  console.log('END OF DONE ANIMATION')
+}
+
 stories.add(
   'default',
   withInfo('')(() => (
     <Loader
       inline={boolean('inline', true)}
       size={number('size', 48)}
+      done={boolean('done', false)}
+      onDoneAnimationEnd={onAnimationEnd}
     />
   )),
 )