+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
-- [Button] API breaking changes (repercussions on ItemChoice and AutoComplete API)
+- [Loader] Now handles the "green check" animation and callback
+- [Button & ItemChoice] API breaking changes, now use a unique `status` instead of multiple props to avoid overlaps of properties and styles
+- [Radio & AutoComplete] Repercussions from ItemChoice API changes
 
 #v0.0.2
 - [Button] Shadowed prop on secondary button